delete + new

+++ views/mahasiswa_mk/delete.php
@@ -1,9 +1,17 @@
 <?php
 	$id = $_GET['id'];
-	$mahasiswa_id = $_GET['mahasiswa_id'];
-
-	$result = mysqli_query($conn, "DELETE FROM mk_mahasiswa WHERE id=$id");
 	
-	header("Location: " . "http://localhost/ilham?page=mahasiswa");
+
+	$sql = "SELECT mahasiswa_id FROM mk_mahasiswa WHERE id=".$id;
+	$result = mysqli_query($conn, $sql);
+	if (mysqli_num_rows($result) > 0) {
+	    while($row = mysqli_fetch_assoc($result)) {
+	    	$mahasiswa_id = $row["mahasiswa_id"];
+			$result = mysqli_query($conn, "DELETE FROM mk_mahasiswa WHERE id=$id");
+			
+			header("Location: " . "http://localhost/ilham?page=mahasiswa_mk&id=".$mahasiswa_id);
+	    }
+	}
+
 
 ?>
